adds position to screenshots renames screenshots to a different name to bypass the 1k limit

+++ src/gui.cpp
@@ -14,12 +14,9 @@ std::string selected_screenshot;
 void render_file_browser()
 {
     if (ImGui::BeginChild("Filesystem##ScreenshotFS", {150, -FLT_MIN}, true)) {
-        for (std::filesystem::directory_iterator dir(Settings::screenshots_path); const auto &entry : dir) {
-            if (entry.is_regular_file()) {
-                if (ImGui::Selectable(entry.path().filename().string().c_str(),
-                                      selected_screenshot == entry.path().filename().string())) {
-                    selected_screenshot = entry.path().filename().string();
-                }
+        for (const auto &[name, path, position] : Settings::screenshots) {
+            if (ImGui::Selectable(name.c_str(), selected_screenshot == name)) {
+                selected_screenshot = name;
             }
         }
         ImGui::EndChild();
@@ -33,8 +30,18 @@ void render_screenshot()
         ImGui::Text("%s", selected_screenshot.c_str());
         ImGui::SameLine(0, 1 * ImGui::GetStyle().ItemSpacing.x);
         if (ImGui::SmallButton("Delete##ScreenshotDelete")) {
+            if (selected_screenshot.empty())
+                return;
             std::filesystem::remove(Settings::screenshots_path.string() + "\\" + selected_screenshot);
+            Settings::screenshots.erase(
+                std::ranges::find(Settings::screenshots, selected_screenshot, &Settings::Screenshot::name));
+            Settings::json_settings[Settings::SCREENSHOTS] = Settings::screenshots;
+            Settings::save(Settings::settings_path);
+            selected_screenshot.clear();
         }
+        const auto screenshot =
+            std::ranges::find(Settings::screenshots, selected_screenshot, &Settings::Screenshot::name);
+        ImGui::Text("Taken at: %f %f", screenshot->position.X, screenshot->position.Y);
         if (const auto pos = selected_screenshot.find('.'); pos != std::string::npos) {
             const std::string identifier = std::string("SCREENSHOTS_IMAGE_").append(selected_screenshot.substr(0, pos));
             if (!textures.contains(identifier)) {
